+++ coroutines/compose/src/commonMain/kotlin/dev/inmo/micro_utils/coroutines/compose/FlowAsListState.kt
@@ -5,11 +5,10 @@ import androidx.compose.runtime.snapshots.SnapshotStateList
 import dev.inmo.micro_utils.common.applyDiff
 import dev.inmo.micro_utils.coroutines.subscribeSafelyWithoutExceptions
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
 
 @Suppress("NOTHING_TO_INLINE")
-inline fun <reified T> StateFlow<List<T>>.toMutableListState(
+inline fun <reified T> StateFlow<List<T>>.asMutableComposeListState(
     scope: CoroutineScope
 ): SnapshotStateList<T> {
     val state = mutableStateListOf(*value.toTypedArray())
@@ -18,4 +17,9 @@ inline fun <reified T> StateFlow<List<T>>.toMutableListState(
     }
     return state
 }
+
+@Suppress("NOTHING_TO_INLINE")
+inline fun <reified T> StateFlow<List<T>>.asComposeList(
+    scope: CoroutineScope
+): List<T> = asMutableComposeListState(scope)